Issue #000 chore: Move deduplication to Validation function
