+++ data-pipeline-flink/druid-events-validator/src/main/resources/druid-events-validator.conf
@@ -29,7 +29,7 @@ schema {
 
 redis {
   database {
-    duplicationstore.id = 3
+    duplicationstore.id = 8
     key.expiry.seconds = 3600
   }
 }
