+++ data-pipeline-flink/druid-events-validator/src/main/scala/org/sunbird/dp/validator/functions/DruidValidatorFunction.scala
@@ -1,35 +1,54 @@
 package org.sunbird.dp.validator.functions
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
+import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.functions.ProcessFunction
 import org.slf4j.LoggerFactory
+import org.sunbird.dp.core.cache.{DedupEngine, RedisConnect}
 import org.sunbird.dp.core.job.{BaseProcessFunction, Metrics}
 import org.sunbird.dp.validator.domain.Event
 import org.sunbird.dp.validator.task.DruidValidatorConfig
 import org.sunbird.dp.validator.util.SchemaValidator
 
-class DruidValidatorFunction(config: DruidValidatorConfig)(implicit val eventTypeInfo: TypeInformation[Event])
+class DruidValidatorFunction(config: DruidValidatorConfig,
+                             @transient var schemaValidator: SchemaValidator = null,
+                             @transient var dedupEngine: DedupEngine = null)
+                            (implicit val eventTypeInfo: TypeInformation[Event])
   extends BaseProcessFunction[Event, Event](config) {
 
   private[this] val logger = LoggerFactory.getLogger(classOf[DruidValidatorFunction])
 
-  lazy val schemaValidator: SchemaValidator = new SchemaValidator(config)
+  override def open(parameters: Configuration): Unit = {
+    super.open(parameters)
+    if (dedupEngine == null) {
+      val redisConnect = new RedisConnect(config.redisHost, config.redisPort, config)
+      dedupEngine = new DedupEngine(redisConnect, config.dedupStore, config.cacheExpirySeconds)
+    }
+    if (schemaValidator == null) {
+      schemaValidator = new SchemaValidator(config)
+    }
+  }
+
+  override def close(): Unit = {
+    super.close()
+    dedupEngine.closeConnectionPool()
+  }
 
   override def metricsList(): List[String] = {
-    List(config.processedMetricsCount, config.validationSuccessMetricsCount,
-      config.validationFailureMetricsCount)
+    List(config.validationSuccessMetricsCount, config.validationFailureMetricsCount) ::: deduplicationMetrics
   }
 
   override def processElement(event: Event,
                               ctx: ProcessFunction[Event, Event]#Context,
                               metrics: Metrics): Unit = {
 
-    metrics.incCounter(config.processedMetricsCount)
     val validationReport = schemaValidator.validate(event)
+
     if (validationReport.isSuccess) {
       event.markValidationSuccess()
       metrics.incCounter(config.validationSuccessMetricsCount)
-      ctx.output(config.validEventOutputTag, event)
+      deDup[Event, Event](event.mid(), event, ctx,
+        config.validEventOutputTag, config.duplicateEventOutputTag, flagName = "dv_duplicate")(dedupEngine, metrics)
     } else {
       val failedErrorMsg = schemaValidator.getInvalidFieldName(validationReport.toString)
       event.markValidationFailure(failedErrorMsg)
